requirements.txt
aiogram==2.23.1
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0

